+++ src/main/java/com/windmeal/chat/service/ChatRoomService.java
@@ -16,16 +16,13 @@ import com.windmeal.order.domain.Order;
 import com.windmeal.order.domain.OrderStatus;
 import com.windmeal.order.exception.OrderAlreadyMatchedException;
 import com.windmeal.order.exception.OrderNotFoundException;
-import com.windmeal.order.repository.OrderRepository;
+import com.windmeal.order.repository.order.OrderRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Slf4j
 @Service
