+++ src/main/java/com/windmeal/order/controller/OrderController.java
@@ -124,4 +124,15 @@ public class OrderController {
   }
 
 
+  @GetMapping("/ordered")
+  @Operation(summary = "내가 주문했던 목록 조회", description = "마이 페이지")
+  @ApiResponses({
+      @ApiResponse(responseCode = "200", description = "OK")
+  })
+  public ResultDataResponseDTO getOwnOrdered(Pageable pageable){
+    Long memberId = SecurityUtil.getCurrentMemberId();
+
+
+    return ResultDataResponseDTO.of(null);
+  }
 }
