+++ src/main/java/com/windmeal/order/repository/order/OrderMenuOptionGroupRepository.java
@@ -1,4 +1,4 @@
-package com.windmeal.order.repository;
+package com.windmeal.order.repository.order;
 
 import com.windmeal.order.domain.Delivery;
 import com.windmeal.order.domain.OrderMenuOptionGroup;
