+++ src/main/java/com/windmeal/order/repository/order/OrderRepository.java
@@ -1,9 +1,6 @@
-package com.windmeal.order.repository;
+package com.windmeal.order.repository.order;
 
 import com.windmeal.order.domain.Order;
-import com.windmeal.store.domain.Category;
-import com.windmeal.store.repository.CategoryCustomRepository;
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long>,OrderCustomRepository
