+++ src/main/java/com/windmeal/order/service/DeliveryService.java
@@ -20,12 +20,11 @@ import com.windmeal.order.exception.DeliverOrdererSameException;
 import com.windmeal.order.exception.DeliveryNotFoundException;
 import com.windmeal.order.exception.OrderAlreadyMatchedException;
 import com.windmeal.order.exception.OrderNotFoundException;
-import com.windmeal.order.repository.DeliveryRepository;
-import com.windmeal.order.repository.OrderCancelRepository;
-import com.windmeal.order.repository.OrderRepository;
+import com.windmeal.order.repository.delivery.DeliveryRepository;
+import com.windmeal.order.repository.order.OrderCancelRepository;
+import com.windmeal.order.repository.order.OrderRepository;
 import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
