+++ src/test/java/com/windmeal/order/repository/OrderCustomRepositoryImplTest.java
@@ -1,6 +1,7 @@
 package com.windmeal.order.repository;
 
 import com.windmeal.IntegrationTestSupport;
+import com.windmeal.order.repository.order.OrderRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
