+++ src/test/java/com/windmeal/order/service/DeliveryServiceTest.java
@@ -21,12 +21,12 @@ import com.windmeal.order.exception.DeliverOrdererSameException;
 import com.windmeal.order.exception.DeliveryNotFoundException;
 import com.windmeal.order.exception.OrderAlreadyMatchedException;
 import com.windmeal.order.exception.OrderNotFoundException;
-import com.windmeal.order.repository.DeliveryRepository;
-import com.windmeal.order.repository.OrderCancelRepository;
-import com.windmeal.order.repository.OrderMenuOptionGroupRepository;
-import com.windmeal.order.repository.OrderMenuOptionSpecificationRepository;
-import com.windmeal.order.repository.OrderMenuRepository;
-import com.windmeal.order.repository.OrderRepository;
+import com.windmeal.order.repository.delivery.DeliveryRepository;
+import com.windmeal.order.repository.order.OrderCancelRepository;
+import com.windmeal.order.repository.order.OrderMenuOptionGroupRepository;
+import com.windmeal.order.repository.order.OrderMenuOptionSpecificationRepository;
+import com.windmeal.order.repository.order.OrderMenuRepository;
+import com.windmeal.order.repository.order.OrderRepository;
 import java.util.ArrayList;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
