+++ src/test/java/com/windmeal/order/service/OrderServiceTest.java
@@ -5,7 +5,6 @@ import static com.windmeal.Fixtures.aOrder;
 import static com.windmeal.Fixtures.aStore;
 import static org.assertj.core.api.Assertions.*;
 
-import com.windmeal.Fixtures;
 import com.windmeal.IntegrationTestSupport;
 import com.windmeal.generic.domain.Money;
 import com.windmeal.member.domain.Member;
@@ -14,7 +13,6 @@ import com.windmeal.member.repository.MemberRepository;
 import com.windmeal.model.place.Place;
 import com.windmeal.model.place.PlaceRepository;
 import com.windmeal.order.domain.Order;
-import com.windmeal.order.domain.OrderMenu;
 import com.windmeal.order.domain.OrderStatus;
 import com.windmeal.order.dto.request.OrderCreateRequest;
 import com.windmeal.order.dto.request.OrderCreateRequest.OrderCreateRequestBuilder;
@@ -29,10 +27,10 @@ import com.windmeal.order.dto.response.OrderCreateResponse;
 import com.windmeal.order.exception.OrderAlreadyMatchedException;
 import com.windmeal.order.exception.OrderNotFoundException;
 import com.windmeal.order.exception.OrdererMissMatchException;
-import com.windmeal.order.repository.OrderMenuOptionGroupRepository;
-import com.windmeal.order.repository.OrderMenuOptionSpecificationRepository;
-import com.windmeal.order.repository.OrderMenuRepository;
-import com.windmeal.order.repository.OrderRepository;
+import com.windmeal.order.repository.order.OrderMenuOptionGroupRepository;
+import com.windmeal.order.repository.order.OrderMenuOptionSpecificationRepository;
+import com.windmeal.order.repository.order.OrderMenuRepository;
+import com.windmeal.order.repository.order.OrderRepository;
 import com.windmeal.store.domain.Menu;
 import com.windmeal.store.domain.Menu.MenuBuilder;
 import com.windmeal.store.domain.OptionGroup;
@@ -41,14 +39,12 @@ import com.windmeal.store.domain.OptionSpecification;
 import com.windmeal.store.domain.OptionSpecification.OptionSpecificationBuilder;
 import com.windmeal.store.domain.Store;
 import com.windmeal.store.domain.Store.StoreBuilder;
-import com.windmeal.store.exception.StoreNotFoundException;
 import com.windmeal.store.repository.MenuRepository;
 import com.windmeal.store.repository.StoreRepository;
 import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
